Fix Goldfish TTY write width to use u8 instead of u32
Each character was being written as a 32-bit value to the PUT_CHAR
register. On big-endian m68k this produces a 4-byte MMIO write per
character. Use a u8 pointer so each byte is written with the correct
access width.

+++ arch/m68k/src/serial.rs
@@ -12,10 +12,10 @@ impl GoldfishSerial {
 
 impl KernelOutput for GoldfishSerial {
     fn write_str(&self, s: &str) {
-        let reg = GOLDFISH_TTY_BASE as *mut u32;
+        let reg = GOLDFISH_TTY_BASE as *mut u8;
         for byte in s.bytes() {
             // SAFETY: Goldfish TTY MMIO address is fixed in the QEMU virt machine
-            unsafe { core::ptr::write_volatile(reg, byte as u32) };
+            unsafe { core::ptr::write_volatile(reg, byte) };
         }
     }
 }
